+++ deepparse/network/bpemb_seq2seq.py
@@ -42,6 +42,7 @@ class BPEmbSeq2SeqModel(Seq2SeqModel):
         verbose: bool = True,
         path_to_retrained_model: Union[str, None] = None,
         pre_trained_weights: bool = True,
+        offline: bool = False,
     ) -> None:
         super().__init__(
             device,
@@ -67,7 +68,7 @@ class BPEmbSeq2SeqModel(Seq2SeqModel):
             model_weights_name = "bpemb"
             if attention_mechanism:
                 model_weights_name += "_attention"
-            self._load_pre_trained_weights(model_weights_name, cache_dir=cache_dir)
+            self._load_pre_trained_weights(model_weights_name, cache_dir=cache_dir, offline=offline)
 
     def forward(
         self,