+++ tests/network/integration/test_integration_seq2seq_model_cpu.py
@@ -9,6 +9,7 @@ from unittest.mock import patch
 
 import torch
 
+from deepparse import CACHE_PATH
 from deepparse.network import Seq2SeqModel
 from ..integration.base import Seq2SeqIntegrationTestCase
 
@@ -127,6 +128,17 @@ class Seq2SeqIntegrationTest(Seq2SeqIntegrationTestCase):
         )
 
         random_mock.assert_not_called()
+
+    @patch("deepparse.network.seq2seq.download_weights")
+    def test_givenAnOfflineSeq2SeqModel_whenInit_thenDontCallDownloadWeights(self, download_weights_mock):
+        # Test if functions latest_version and download_weights
+
+        a_model_type = "fasttext"
+
+        default_cache = CACHE_PATH
+        self.pre_trained_seq2seq_model._load_pre_trained_weights(a_model_type, cache_dir=default_cache, offline=True)
+
+        download_weights_mock.assert_not_called()
 
 
 if __name__ == "__main__":