+++ tests/vectorizer/test_train_vectorizer.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 from unittest.mock import MagicMock
 
-from deepparse.data_error import DataError
+from deepparse.errors.data_error import DataError
 from deepparse.embeddings_models import EmbeddingsModel
 from deepparse.vectorizer import TrainVectorizer, BPEmbVectorizer